requirements.txt
Flask~=1.1.2
numpy==1.19.2
pandas==1.1.5
scikit-learn~=0.23.2
gunicorn==20.0.4
requests==2.25.1
pymongo==3.11.2

tensorflow-cpu ~=2.4.0
dnspython==2.1.0